Comparaison token userId pour suppression et modification de sauce

+++ controllers/sauce.js
@@ -41,12 +41,17 @@ exports.modifySauce = (req, res, next) => {
     }
     Sauce.findOne({ _id: req.params.id })
         .then(sauce => {
-            const filename = sauce.imageUrl.split('/images/')[1];
-            fs.unlink('images/' + filename, () => {
-                Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-                    .then(() => res.status(200).json({ message: "Sauce modifiée" }))
-                    .catch(error => res.status(400).json({ error }));
-            })
+            if (sauce.userId !== req.auth.userId) {
+                res.status(403).json({ message: "Modification non autorisée" })
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink('images/' + filename, () => {
+                    Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                        .then(() => res.status(200).json({ message: "Sauce modifiée" }))
+                        .catch(error => res.status(400).json({ error }));
+                })
+            }
+
         })
         .catch(error => res.status(404).json({ error }));
 
@@ -57,12 +62,17 @@ exports.modifySauce = (req, res, next) => {
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
         .then(sauce => {
-            const filename = sauce.imageUrl.split('/images/')[1];
-            fs.unlink('images/' + filename, () => {
-                Sauce.deleteOne({ _id: req.params.id })
-                    .then(() => res.status(200).json({ message: "Sauce supprimée" }))
-                    .catch(error => res.status(400).json({ error }));
-            })
+            if (sauce.userId !== req.auth.userId) {
+                res.status(403).json({ message: "Suppression non autorisée" })
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink('images/' + filename, () => {
+                    Sauce.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: "Sauce supprimée" }))
+                        .catch(error => res.status(400).json({ error }));
+                })
+            }
+
         })
         .catch(error => res.status(404).json({ error }));
 }
